refactor: add "@test" path mapping
