+++ projects/ngx-meta/src/open-graph/src/open-graph-image-metadata-provider.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing'
 import { MockProviders } from 'ng-mocks'
 import { MetadataSetter, NgxMetaMetaService } from '@davidlj95/ngx-meta/core'
-import { enableAutoSpy } from '../../__tests__/enable-auto-spy'
+import { enableAutoSpy } from '@/ngx-meta/test/enable-auto-spy'
 import { OpenGraphImage } from './open-graph-image'
 import { OpenGraph } from './open-graph'
 import { __OPEN_GRAPH_IMAGE_SETTER_FACTORY } from './open-graph-image-metadata-provider'
