+++ projects/ngx-meta/src/standard/src/standard-generator-metadata-provider.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing'
 import { MockProvider } from 'ng-mocks'
-import { enableAutoSpy } from '../../__tests__/enable-auto-spy'
+import { enableAutoSpy } from '@/ngx-meta/test/enable-auto-spy'
 import { MetadataSetter, NgxMetaMetaService } from '@davidlj95/ngx-meta/core'
 import { VERSION } from '@angular/core'
 import { Standard } from './standard'
